added commands and filters to services
