+++ src/main/java/ro/uvt/info/controllers/BooksController.java
@@ -9,6 +9,7 @@ import ro.uvt.info.models.MyPair;
 import ro.uvt.info.services.BookRepository;
 import ro.uvt.info.services.BookStatistics;
 import ro.uvt.info.services.CommandExecutor;
+import ro.uvt.info.services.commands.*;
 
 import java.util.List;
