--- src/main/java/ro/uvt/info/controllers/Filters/RequestLoggingFilter.java
@@ -1,27 +1,0 @@
-package ro.uvt.info.controllers.Filters;
-
-import jakarta.servlet.*;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.core.annotation.Order;
-import org.springframework.stereotype.Component;
-
-import java.io.IOException;
-
-@Component
-@Order(1)
-public class RequestLoggingFilter implements Filter {
-    @Override
-    public void doFilter(
-            ServletRequest request,
-            ServletResponse response,
-            FilterChain chain) throws IOException, ServletException {
-
-        HttpServletRequest req = (HttpServletRequest) request;
-        HttpServletResponse res = (HttpServletResponse) response;
-        System.out.printf("Logging Request  %s : %s \n", req.getMethod(),
-                req.getRequestURI());
-        chain.doFilter(request, response);
-    }
-
-}
