--- src/main/java/ro/uvt/info/controllers/commands/Command.java
@@ -1,7 +1,0 @@
-package ro.uvt.info.controllers.commands;
-
-public interface Command<TReturn, TCommandContext> {
-    public TReturn execute();
-    public default void setCommandContext(TCommandContext commandContext){}
-    public Command<TReturn, TCommandContext> getClone();
-}
