+++ src/main/java/ro/uvt/info/services/CommandExecutor.java
@@ -3,8 +3,7 @@ package ro.uvt.info.services;
 import org.springframework.beans.factory.config.ConfigurableBeanFactory;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
-import ro.uvt.info.controllers.commands.Command;
-import ro.uvt.info.models.MyPair;
+import ro.uvt.info.services.commands.Command;
 
 import java.util.*;
 import java.util.concurrent.ExecutorService;
