+++ src/main/java/ro/uvt/info/services/commands/Command.java
@@ -1,0 +1,7 @@
+package ro.uvt.info.services.commands;
+
+public interface Command<TReturn, TCommandContext> {
+    public TReturn execute();
+    public default void setCommandContext(TCommandContext commandContext){}
+    public Command<TReturn, TCommandContext> getClone();
+}
